Clear pids after each round so perf failures don't occur (LP:#1494226)
Not clearing the pid and checking for a cleared pid in the perf stats
means that perf open fails (because the pid does not exist from a previous
round of testing) causing PERF_INVALID to be set in the counter which in
turn leads to huge bogus perf counter stats.  After each stressor process
termination we should zero the pid and don't run perf stats on zero'd pids
on subsequent test rounds.

Signed-off-by: Colin Ian King <colin.king@canonical.com>